Bugs Fixed: - Profile cancel null data - Updating profile wrong data username format for lender - Removed all of onBackPressed - Payment incorrect remaining payment - Dashboard now stops including posponed/on list borrower users
Improvements
- User Interface
- Total now includes interest
- Remaining now includes interest
- All data list now automatically udpates each confirms
- All data now prevents null
- Regex improvement
- Removed postdelayed handler for better performance

+++ Android System/app/src/main/java/com/example/model/applyModel.java
@@ -5,9 +5,9 @@ import android.graphics.Bitmap;
 public class applyModel {
 
     String name, email;
-    int amount;
+    double amount;
     Bitmap img;
-    public applyModel(String name, String email, int amount, Bitmap img) {
+    public applyModel(String name, String email, double amount, Bitmap img) {
         this.name = name;
         this.email = email;
         this.amount = amount;
@@ -18,7 +18,7 @@ public class applyModel {
         return name;
     }
 
-    public int getAmount() {
+    public double getAmount() {
         return amount;
     }
